PDFExtractionNew/src/requirement.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
blis==0.4.1
boto==2.49.0
boto3==1.14.7
botocore==1.17.7
bs4==0.0.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
Cython==0.29.14
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
en-core-web-sm==2.3.0
extract-msg==0.23.1
futures==3.1.1
gensim==3.8.3
idna==2.9
IMAPClient==2.1.0
importlib-metadata==1.6.1
jmespath==0.10.0
joblib==0.14.1
lxml==4.5.1
murmurhash==1.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.20
nltk==3.5
numpy==1.18.5
olefile==0.46
pandas==0.25.3
pdfminer.six==20181108
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
pycryptodome==3.9.7
PyPDF2==1.26.0
pytesseract==0.3.4
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
regex==2020.6.8
requests==2.24.0
s3transfer==0.3.3
scipy==1.4.1
six==1.12.0
smart-open==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.0
SpeechRecognition==3.8.1
srsly==1.0.2
textract==1.6.3
thinc==7.4.1
tqdm==4.46.1
tzlocal==1.5.1
urllib3==1.25.9
wasabi==0.7.0
xlrd==1.2.0
XlsxWriter==1.2.9
zipp==1.2.0
